+++ OneBlink.SDK/models/Form.cs
@@ -24,7 +24,9 @@ namespace OneBlink.SDK.Model
             bool isInfoPage = false,
             List<string> tags = default(List<string>),
             DateTime? publishStartDate = null,
-            DateTime? publishEndDate = null
+            DateTime? publishEndDate = null,
+            string cancelAction = "BACK",
+            string cancelRedirectUrl = null
             )
         {
             if (id.HasValue)
@@ -80,8 +82,10 @@ namespace OneBlink.SDK.Model
             {
                 this.publishEndDate = publishEndDate.Value;
             }
+            this.cancelAction = cancelAction;
+            this.cancelRedirectUrl = cancelRedirectUrl;
         }
-        private string[] AllowedPostSubmissionActions = new string[] { "URL", "CLOSE", "FORMS_LIBRARY" };
+        private string[] AllowedPostSubmissionActions = new string[] { "BACK", "URL", "CLOSE", "FORMS_LIBRARY" };
         public long id { get; set; }
         public string name { get; set; }
         public string description { get; set; }
@@ -115,5 +119,22 @@ namespace OneBlink.SDK.Model
         public List<string> tags { get; set; }
         public DateTime? publishStartDate { get; set; }
         public DateTime? publishEndDate { get; set; }
+        private string _CancelAction;
+        public string cancelAction
+        {
+            get
+            {
+                return _CancelAction;
+            }
+            set
+            {
+                if (!AllowedPostSubmissionActions.Any(x => x == value))
+                {
+                    throw new ArgumentException(value + " not a valid cancel action");
+                }
+                _CancelAction = value;
+            }
+        }
+        public string cancelRedirectUrl { get; set; }
     }
 }